refactor: move Basket to core.domain, fix Coupon.id generation strategy to match AUTO_INCREMENT in data.sql
