+++ review/src/test/java/schwarz/jobs/interview/coupon/core/domain/BasketValidationTest.java
@@ -1,10 +1,11 @@
-package schwarz.jobs.interview.coupon.core.services.model;
+package schwarz.jobs.interview.coupon.core.domain;
 
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import schwarz.jobs.interview.coupon.core.domain.Basket;
 
 import java.math.BigDecimal;
 import java.util.Set;
